+++ src/pvoutput.cpp
@@ -12,12 +12,12 @@ void Pulse::setPVOutput(char const* t_apikey, char const* t_sysid,
   char const* t_url, int const& t_interval)
 {
   if ((!t_apikey) || (!t_sysid) || (!t_url) || (!t_interval)) {
-    cout << "Upload to PVOutput.org disabled" << endl;
+    cout << "Upload to PVOutput disabled" << endl;
     return;
   } 
   
   if (!(t_interval == 5 || t_interval == 10 || t_interval == 15)) {
-    throw runtime_error(string("Upload to PVOutput.org every ")
+    throw runtime_error(string("Upload to PVOutput every ")
       + to_string(t_interval) + " minutes not supported");
   }
   
@@ -28,7 +28,7 @@ void Pulse::setPVOutput(char const* t_apikey, char const* t_sysid,
   m_interval = t_interval;
     
   if (m_debug) {
-    cout << "Upload to PVOutput.org every " << t_interval
+    cout << "Upload to PVOutput every " << t_interval
       << " minutes" << endl;
   }
 }
@@ -88,14 +88,15 @@ void Pulse::uploadXport(void) const
   curl_easy_setopt(easyhandle, CURLOPT_WRITEDATA, &read_buffer);
 
   CURLcode res = curl_easy_perform(easyhandle);
-  if (res != CURLE_OK) {
-    throw runtime_error(string("Curl_easy_perform() failed: ")
-      + curl_easy_strerror(res));
-  }
- 
   curl_easy_cleanup(easyhandle);
   curl_slist_free_all(headers);
 
+  if (res != CURLE_OK) {
+    cout << "Upload to PVOutput failed: " 
+         << curl_easy_strerror(res) << endl;
+    return;
+  }
+ 
   char buffer[32] = {0};
   strftime(buffer, 31, "%F %T", tm);